Rename scenario files to .scenario.js and fix stale title/description
Both files were still carrying the old title, description, and hardcoded region prefix from before they were split apart, so the cli:seed picker showed misleading and duplicate entries.
