+++ tests/internal/return-versions/cancel-returns.spec.js
@@ -1,4 +1,4 @@
-import scenarioData from '../../support/scenarios/unregistered-licence-with-two-purposes-and-requirements.js'
+import scenarioData from '../../support/scenarios/unregistered-licence-with-two-purposes-and-requirements.scenario.js'
 import { test, expect } from '../../support/fixtures.js'
 
 const scenario = scenarioData()
